BACKEND_API/requirements.txt
fastapi 
uvicorn
fastapi-cache2
psycopg2-binary
pydantic
python-jose
passlib[bcrypt]
jinja2
